# docs/swagger.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoGoStudy API",
    "version": "1.0.0",
    "description": "GoGoStudy 인증/사용자/스터디/출석 API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "path": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "enum": [
              "ACCOUNT_INACTIVE",
              "ALREADY_JOINED",
              "EMAIL_TAKEN",
              "FORBIDDEN",
              "GOOGLE_AUTH_NOT_CONFIGURED",
              "INVALID_CREDENTIALS",
              "INVALID_DATE",
              "INVALID_GOOGLE_TOKEN",
              "INVALID_ID",
              "INVALID_MEMBER_STATUS",
              "INVALID_OPERATION",
              "INVALID_PASSWORD",
              "INVALID_PAYLOAD",
              "INVALID_ROLE",
              "INVALID_STATUS",
              "INVALID_STUDY",
              "INVALID_TOKEN",
              "INVALID_USER_ID",
              "MEMBER_NOT_FOUND",
              "MEMBERSHIP_NOT_FOUND",
              "NOT_A_MEMBER",
              "PASSWORD_TOO_SHORT",
              "REFRESH_NOT_FOUND",
              "SESSION_NOT_FOUND",
              "STUDY_FULL",
              "STUDY_NOT_FOUND",
              "TOKEN_REVOKED",
              "UNAUTHORIZED",
              "UNKNOWN_ERROR",
              "USER_NOT_FOUND",
              "INTERNAL_ERROR"
            ],
            "description": "백엔드에서 사용하는 에러 코드"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "timestamp",
          "path",
          "status",
          "code",
          "message"
        ],
        "example": {
          "timestamp": "2024-12-24T00:00:00.000Z",
          "path": "/auth/login",
          "status": 401,
          "code": "INVALID_CREDENTIALS",
          "message": "Invalid credentials",
          "details": null
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Study": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "maxMembers": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "leaderId": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AttendanceRecord": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "sessionId": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 요청",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "찾을 수 없음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "서버 오류",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 429
                },
                "code": {
                  "type": "string",
                  "example": "RATE_LIMITED"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Service unavailable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "service": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/studies": {
      "get": {
        "tags": [
          "Studies"
        ],
        "summary": "스터디 목록",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "?? ???"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "?? ???(q? ??)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "??? (0?? ??)"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "??? ?? (?? 20, ?? 50)"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "??: field,ASC|DESC ?) createdAt,DESC"
          }
        ],
        "responses": {
          "200": {
            "description": "?? ?? ??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Study"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "totalElements": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "sort": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/인증 실패"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Studies"
        ],
        "summary": "스터디 생성",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "maxMembers": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "스터디 생성 완료",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "study": {
                      "$ref": "#/components/schemas/Study"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/인증 실패"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/admin/users/{id}/attendance": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "사용자 출석 기록 조회",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "출석 이력",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    },
                    "records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AttendanceRecord"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/인증 실패"
          },
          "403": {
            "$ref": "#/components/responses/접근 거부"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/admin/studies": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "스터디 목록 (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "?? ???"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "?? ???(q? ??)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "??? (0?? ??)"
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "??? ?? (?? 20, ?? 50)"
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "??: field,ASC|DESC ?) createdAt,DESC"
          }
        ],
        "responses": {
          "200": {
            "description": "?? ?? ??",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Study"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "totalElements": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "sort": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/인증 실패"
          },
          "403": {
            "$ref": "#/components/responses/접근 거부"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/admin/stats/overview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "플랫폼 통계 요약",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Overview stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "active": {
                          "type": "integer"
                        },
                        "admins": {
                          "type": "integer"
                        }
                      }
                    },
                    "studies": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "recruiting": {
                          "type": "integer"
                        }
                      }
                    },
                    "attendance": {
                      "type": "object",
                      "properties": {
                        "sessionsTotal": {
                          "type": "integer"
                        },
                        "sessionsToday": {
                          "type": "integer"
                        },
                        "records": {
                          "type": "integer"
                        },
                        "pendingMembers": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/인증 실패"
          },
          "403": {
            "$ref": "#/components/responses/접근 거부"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  }
}